requirements.txt
requests
python-firebase
numpy
matplotlib
h5py
scikit-image